Consolidate conversion layer types and functions
- Move ProviderData, ToolCall, ToolResult types to backend/tool/types
- Remove duplicate ConvertMemoryMessageToProto from backend/agent
- Replace simplified api/conv/message.go with comprehensive implementation
- Simplify shared/conv to only utility functions (Ptr, FromPtr, ErrorToString)

This consolidation establishes clear ownership:
- backend/tool/types: persistence types for tool calls/results
- backend/api/conv: Memory↔Proto conversions (API layer)
- backend/agent: Memory↔Model conversions (LLM provider layer)
